pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-taxsystem"
description = "Tax System module for Alliance Auth"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Geuthur", email = "devgeuthur@gmail.com" },
]
requires-python = ">=3.10"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.10,<5",
    "django-esi>=8,<9",
    "django-eveuniverse>=1.6",
    "django-ninja>=1.5,<2",
]
optional-dependencies.tests-allianceauth-latest = [
    "coverage",
    "factory-boy",
]
urls.Changelog = "https://github.com/Geuthur/aa-taxsystem/blob/master/CHANGELOG.md"
urls.Homepage = "https://github.com/Geuthur/aa-taxsystem"
urls.Source = "https://github.com/Geuthur/aa-taxsystem"
urls.Tracker = "https://github.com/Geuthur/aa-taxsystem/issues"

[tool.hatch.build]
include = [
    "/taxsystem",
]

[tool.hatch.version]
path = "taxsystem/__init__.py"
